requirements.txt
selenium
webdriver-manager
python-dotenv
PyYAML
requests
